+++ src/routes/API/user.js
@@ -5,7 +5,7 @@ import {
   sendOtp,
   loginUser,
   deleteUser,
-} from '../../controllers/userController';
+} from '../../controllers/userController001';
 import verifVendorJWT from '../../middleware/verifyJWT';
 
 const usersRouter = express.Router();
